// x-pack/platform/plugins/shared/ai_infra/product_doc_base/server/services/doc_manager/types.d.ts
export interface SecurityLabsStatusResponse {
    status: InstallationStatus;
    version?: string;
    latestVersion?: string;
    isUpdateAvailable?: boolean;
    failureReason?: string;
}